+++ js/ai/qlearn/changelog.md
@@ -5,3 +5,4 @@
  * Start a changelog
  * Change the learn calculation
  * Up to ecma2021
+ * rename actionNames into previousActions inside learn